pyproject.toml
[project]
name = "vn-law-chatbot"
version = "0.1.0"
description = "Vietnamese civil law chatbot using RAG (local-first)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "chromadb>=0.5.5",
  "fastapi>=0.115.0",
  "httpx>=0.27.2",
  "langchain>=0.3.0",
  "langchain-community>=0.3.0",
  "loguru>=0.7.2",
  "pydantic>=2.9.2",
  "pymongo>=4.9.1",
  "pyvi>=0.1.1",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "ragas>=0.2.0",
  "requests>=2.32.3",
  "sentence-transformers>=3.1.1",
  "streamlit>=1.39.0",
  "transformers>=4.45.0",
  "underthesea>=6.8.4",
  "uvicorn>=0.31.0",
  "sentencepiece>=0.2.1",
  "protobuf>=6.33.5",
]

[dependency-groups]
dev = [
  "pytest>=8.3.3",
  "ruff>=0.6.9",
]

[project.scripts]
crawl-legal-docs = "scripts.crawl_legal_docs:main"
crawl-qa-dataset = "scripts.crawl_qa_dataset:main"
preprocess-qa-dataset = "scripts.preprocess_qa_dataset:main"
build-embeddings = "scripts.build_embeddings:main"
build-qa-embeddings = "scripts.build_qa_embeddings:main"
evaluate-rag = "scripts.evaluate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
